requirements.txt
fastapi[Standard]
langchain
langchain_community
langchain_text_splitters
langchain_huggingface
langchain_groq
langchain_core
torch
transformers
python_dotenv
uvicorn
pypdf
chromadb
pdfplumber
pdf2image
pytesseract